# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/columns/{columnId}/tasks": {
      "post": {
        "operationId": "TaskController_create",
        "parameters": [
          {
            "name": "columnId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDtoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "payload": {
                      "id": 11,
                      "images": null,
                      "name": "name",
                      "description": null,
                      "deadline": "2024-03-05T14:49:39.907Z",
                      "owner_id": 1,
                      "column_id": 5,
                      "statusId": 1,
                      "timestamps": "2024-03-04T14:49:39.907Z",
                      "recepient_id": 1,
                      "photos": "http://photo-link"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "name should not be empty",
                      "name must be shorter than or equal to 255 characters",
                      "name must be a string"
                    ],
                    "error": "Bad request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Column not found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "TaskController_get",
        "parameters": [
          {
            "name": "columnId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "payload": [
                      {
                        "id": 1,
                        "images": null,
                        "name": "task_name",
                        "description": null,
                        "deadline": "2024-02-28T20:21:21.168Z",
                        "owner_id": 1,
                        "column_id": 4,
                        "statusId": 1,
                        "timestamps": "2024-02-26T14:41:42.307Z",
                        "photos": "http://photo-link"
                      },
                      {
                        "id": 2,
                        "images": null,
                        "name": "task_name",
                        "description": null,
                        "deadline": "2024-02-28T20:21:21.168Z",
                        "owner_id": 1,
                        "column_id": 4,
                        "statusId": 1,
                        "timestamps": "2024-02-26T14:41:47.734Z",
                        "photos": "http://photo-link"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Column not found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/columns/{columnId}/tasks/{taskId}": {
      "patch": {
        "operationId": "TaskController_update",
        "parameters": [
          {
            "name": "columnId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDtoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "payload": [
                      {
                        "id": 1,
                        "images": null,
                        "name": "task_name",
                        "description": null,
                        "deadline": "2024-02-28T20:21:21.168Z",
                        "owner_id": 1,
                        "column_id": 4,
                        "statusId": 1,
                        "timestamps": "2024-02-26T14:41:42.307Z",
                        "photos": "http://photo-link"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Column not found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "delete": {
        "operationId": "TaskController_delete",
        "parameters": [
          {
            "name": "columnId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task was been deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "payload": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Column not found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "task-hub-api",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateTaskDtoRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "The description of the task.",
            "example": "Some kind of description"
          },
          "name": {
            "type": "string",
            "description": "The name of the task.",
            "example": "Name of task"
          },
          "deadline": {
            "format": "date-time",
            "type": "string",
            "description": "Deadline date",
            "example": "2024-02-24T20:21:21.168Z"
          },
          "statusId": {
            "type": "object",
            "description": "Status id",
            "example": "1"
          },
          "recepientId": {
            "type": "object",
            "description": "Id of user which assigned to task",
            "example": "1"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTaskDtoRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "The description of the task.",
            "example": "Some kind of description"
          },
          "name": {
            "type": "string",
            "description": "The name of the task.",
            "example": "Name of task"
          },
          "deadline": {
            "format": "date-time",
            "type": "string",
            "description": "Deadline date",
            "example": "2024-02-24T20:21:21.168Z"
          },
          "statusId": {
            "type": "object",
            "description": "Status id",
            "example": "1"
          },
          "deletedImagesId": {
            "type": "object",
            "description": "deleted photos id",
            "example": [
              0,
              4
            ]
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}